% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Behavior.R
\name{C_BehBoundFirstOrder}
\alias{C_BehBoundFirstOrder}
\title{A CHSM Behavior for simulating first order solute reactions}
\description{
An R6 class mapping to a C++ composite Hierarchy State Machine object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Behavior]{rchsm::C_Behavior}} -> \code{C_BehBoundFirstOrder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_BehBoundFirstOrder-new}{\code{C_BehBoundFirstOrder$new()}}
\item \href{#method-C_BehBoundFirstOrder-createVariables}{\code{C_BehBoundFirstOrder$createVariables()}}
\item \href{#method-C_BehBoundFirstOrder-clone}{\code{C_BehBoundFirstOrder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehBoundFirstOrder-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehBoundFirstOrder-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehBoundFirstOrder$new(
  soluteName,
  className = "BehBoundFirstOrder",
  regFinalizer = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soluteName}}{Character string with the name of the solute}

\item{\code{className}}{The name of the associated C class}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered (TRUE)
or not (FALSE).
Defaults to TRUE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehBoundFirstOrder-createVariables"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehBoundFirstOrder-createVariables}{}}}
\subsection{Method \code{createVariables()}}{
Create the variables controlling the behavior in the provided holon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehBoundFirstOrder$createVariables(matrix, holon, initRate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{The matrix object where the variables should be installed}

\item{\code{holon}}{The holon object where the variables should be installed}

\item{\code{initRate}}{Numeric for initial rate of reaction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehBoundFirstOrder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehBoundFirstOrder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehBoundFirstOrder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
