% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Behavior.R
\name{C_BehCellSolute}
\alias{C_BehCellSolute}
\title{A CHSM Behavior for simulating solute concentrations in a cell}
\description{
An R6 class mapping to a C++ composite Hierarchy State Machine object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Behavior]{rchsm::C_Behavior}} -> \code{C_BehCellSolute}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_BehCellSolute-new}{\code{C_BehCellSolute$new()}}
\item \href{#method-C_BehCellSolute-createVariables}{\code{C_BehCellSolute$createVariables()}}
\item \href{#method-C_BehCellSolute-clone}{\code{C_BehCellSolute$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehCellSolute-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehCellSolute-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehCellSolute$new(
  soluteName,
  className = "BehCellSolute",
  regFinalizer = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{soluteName}}{Character string with the name of the solute}

\item{\code{className}}{The name of the associated C class}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered (TRUE)
or not (FALSE).
Defaults to TRUE.
See super class C_Object for more information.}

\item{\code{mfDouble}}{(Optional) Memory factory used for the state phase values.
Default value is NULL, which may not be compatible with solvers
that require memory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehCellSolute-createVariables"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehCellSolute-createVariables}{}}}
\subsection{Method \code{createVariables()}}{
Create the variables controlling the behavior in the provided holon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehCellSolute$createVariables(
  matrix,
  holon,
  timeHolon,
  timeStepName,
  initConc,
  mfDouble = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{The matrix object where the variables should be installed.}

\item{\code{holon}}{The holon object where the variables should be installed.}

\item{\code{timeHolon}}{The holon containing the time step variable.}

\item{\code{timeStepName}}{The name of the time step variable.}

\item{\code{initConc}}{Numeric for initial concentration of solute.}

\item{\code{mfDouble}}{(Optional) The memory factory for constructing the appropriate
memory for the concentration value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehCellSolute-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehCellSolute-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehCellSolute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
