% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Behavior.R
\name{C_BehCellTime}
\alias{C_BehCellTime}
\title{A CHSM Behavior for simulating time}
\description{
An R6 class mapping to a C++ composite Hierarchy State Machine object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Behavior]{rchsm::C_Behavior}} -> \code{C_BehCellTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_BehCellTime-new}{\code{C_BehCellTime$new()}}
\item \href{#method-C_BehCellTime-createVariables}{\code{C_BehCellTime$createVariables()}}
\item \href{#method-C_BehCellTime-clone}{\code{C_BehCellTime$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehCellTime-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehCellTime-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehCellTime$new(className = "BehCellTime", regFinalizer = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{(Optional) The name of the class for the behavior.
Defaults to "BehCellTime".}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered (TRUE)
or not (FALSE).
Defaults to TRUE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehCellTime-createVariables"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehCellTime-createVariables}{}}}
\subsection{Method \code{createVariables()}}{
Create the variables controlling the behavior in the provided holon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehCellTime$createVariables(
  matrix,
  holon,
  initTime,
  initIteration,
  initTimeValid
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{The matrix object where the variables should be installed}

\item{\code{holon}}{The holon object where the variables should be installed}

\item{\code{initTime}}{Numeric for initial value of time}

\item{\code{initIteration}}{Integer for initial value of the iteration counter}

\item{\code{initTimeValid}}{Logical for initial value of the valid time indicator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehCellTime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehCellTime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehCellTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
