% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_BehDOConcCell.R
\name{C_BehDOConcCell}
\alias{C_BehDOConcCell}
\title{BehDOConcCell Object}
\description{
An R6 class mapping to a C++ BehDOConcCell object

  Implementation is provided by C++ objects.
}
\section{Super class}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{C_BehDOConcCell}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_BehDOConcCell-new}{\code{C_BehDOConcCell$new()}}
\item \href{#method-C_BehDOConcCell-createVariables}{\code{C_BehDOConcCell$createVariables()}}
\item \href{#method-C_BehDOConcCell-clone}{\code{C_BehDOConcCell$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="finalize"><a href='../../rchsm/html/C_Object.html#method-C_Object-finalize'><code>rchsm::C_Object$finalize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehDOConcCell-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehDOConcCell-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehDOConcCell$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the cell}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehDOConcCell-createVariables"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehDOConcCell-createVariables}{}}}
\subsection{Method \code{createVariables()}}{
Create the variables controlling the behavior in the provided holon
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehDOConcCell$createVariables(holon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{holon}}{The holon object where the variables should be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_BehDOConcCell-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_BehDOConcCell-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_BehDOConcCell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
