% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_DepManInstallOrder.R
\name{C_DepManInstallOrder}
\alias{C_DepManInstallOrder}
\title{CHSM Install order dependency manager object}
\description{
An R reference to a install order dependency manager.

  Implementation is provided by C++ objects.
}
\section{Super class}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{C_DepManInstallOrder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_DepManInstallOrder-new}{\code{C_DepManInstallOrder$new()}}
\item \href{#method-C_DepManInstallOrder-clone}{\code{C_DepManInstallOrder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_DepManInstallOrder-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_DepManInstallOrder-new}{}}}
\subsection{Method \code{new()}}{
Create an instance of the dependency manager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_DepManInstallOrder$new(
  numPhases,
  className = "DepManInstallOrder",
  regFinalizer = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numPhases}}{Number of phases for dynamic value updates}

\item{\code{className}}{(Optional) Character string with the name of the class.
Defaults to "DepManInstallOrder".}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered (TRUE)
or not (FALSE).
Defaults to FALSE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_DepManInstallOrder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_DepManInstallOrder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_DepManInstallOrder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
