% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Matrix.R
\name{C_Matrix}
\alias{C_Matrix}
\title{CHSM Matrix Object}
\description{
A Composite Hierarchy State Machine matrix value.

  Implementation is provided by C++ objects.
}
\section{Super class}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{C_Matrix}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_Matrix-new}{\code{C_Matrix$new()}}
\item \href{#method-C_Matrix-createBound}{\code{C_Matrix$createBound()}}
\item \href{#method-C_Matrix-createCell}{\code{C_Matrix$createCell()}}
\item \href{#method-C_Matrix-createVariable}{\code{C_Matrix$createVariable()}}
\item \href{#method-C_Matrix-installSolver}{\code{C_Matrix$installSolver()}}
\item \href{#method-C_Matrix-clone}{\code{C_Matrix$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Matrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_Matrix-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Matrix$new(depManager, className = "Matrix", regFinalizer = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depManager}}{The dependency manager C object the matrix should use}

\item{\code{className}}{(Optional) Name of the class used in accessing c++ functions.
Default value is "Matrix"}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be 
registered (TRUE) or not (FALSE).
Defaults to TRUE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Matrix-createBound"></a>}}
\if{latex}{\out{\hypertarget{method-C_Matrix-createBound}{}}}
\subsection{Method \code{createBound()}}{
Creates a new bound in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Matrix$createBound(name, cellFrom, cellTo, holon = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the bound to be created}

\item{\code{cellFrom}}{R6 bound object for the from cell}

\item{\code{cellTo}}{R6 bound object for the to cell}

\item{\code{holon}}{(Optional) R6 holon object where the bound should be created.
Default NULL value results in creation of the bound in the model holon}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 bound object mapped to the C++ bound object that was created.
  The function will cause a fatal error in R if the c object returns
    an error message.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Matrix-createCell"></a>}}
\if{latex}{\out{\hypertarget{method-C_Matrix-createCell}{}}}
\subsection{Method \code{createCell()}}{
Creates a new cell in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Matrix$createCell(name, holon = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the cell to be created}

\item{\code{holon}}{(Optional) R6 holon object where the cell should be created.
Default NULL value results in creation of the cell in the model holon}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 cell object mapped to the C++ cell object that was created.
  The function will cause a fatal error in R if the c object returns
    an error message.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Matrix-createVariable"></a>}}
\if{latex}{\out{\hypertarget{method-C_Matrix-createVariable}{}}}
\subsection{Method \code{createVariable()}}{
Creates a new variable in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Matrix$createVariable(name, value, holon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string with the name of the variable}

\item{\code{value}}{Value object of the new variable}

\item{\code{holon}}{Holon object containing the new variable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
External pointer to the variable that was created
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Matrix-installSolver"></a>}}
\if{latex}{\out{\hypertarget{method-C_Matrix-installSolver}{}}}
\subsection{Method \code{installSolver()}}{
Install a solver for the matrix updater
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Matrix$installSolver(solver)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solver}}{R6 Solver object to use in the matrix updater}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns a NULL value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Matrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_Matrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Matrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
