% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_MemoryFactory.R
\name{C_MemoryFactory}
\alias{C_MemoryFactory}
\title{A memory factory for a value}
\description{
An R6 class mapping to a C++ Value object

  Implementation is provided by C++ objects.
}
\section{Super class}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{C_MemoryFactory}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_MemoryFactory-new}{\code{C_MemoryFactory$new()}}
\item \href{#method-C_MemoryFactory-clone}{\code{C_MemoryFactory$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_MemoryFactory-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_MemoryFactory-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a memory factory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_MemoryFactory$new(className, size, ..., regFinalizer = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{The name of the C class for the memory factory}

\item{\code{size}}{The size of the memory the factory should create}

\item{\code{...}}{Generic arguments to be passed on to the constructor}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be 
registered (TRUE) or not (FALSE).
Defaults to FALSE
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_MemoryFactory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_MemoryFactory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_MemoryFactory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
