% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Solver.R
\name{C_Solver}
\alias{C_Solver}
\title{CHSM Solver object}
\description{
An R reference to a solver object

  Implementation is provided by C++ objects.
}
\section{Super class}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{C_Solver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_Solver-new}{\code{C_Solver$new()}}
\item \href{#method-C_Solver-clone}{\code{C_Solver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Solver-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_Solver-new}{}}}
\subsection{Method \code{new()}}{
Create an instance of the solver
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Solver$new(className, timeStep, regFinalizer = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{Character string with the name of the class.}

\item{\code{timeStep}}{Variable object for the time step}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered 
(TRUE) or not (FALSE).
Defaults to FALSE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Solver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_Solver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Solver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
