% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Value.R
\name{C_ValueDouble}
\alias{C_ValueDouble}
\title{A double precision floating point value}
\description{
An R6 class mapping to a C++ Value object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Value]{rchsm::C_Value}} -> \code{C_ValueDouble}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_ValueDouble-new}{\code{C_ValueDouble$new()}}
\item \href{#method-C_ValueDouble-clone}{\code{C_ValueDouble$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ValueDouble-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_ValueDouble-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a ValueDouble class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ValueDouble$new(initValue, className = "ValueDouble", regFinalizer = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initValue}}{A numeric for the initial value of the variable}

\item{\code{className}}{(Optional) Character string with the name of the type of value 
to be created.
Default value is "ValueDouble".}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered 
(TRUE) or not (FALSE).
Defaults to FALSE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ValueDouble-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_ValueDouble-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ValueDouble$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
