% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Bound.R
\name{C_Bound}
\alias{C_Bound}
\title{Bound Object}
\description{
An R6 class mapping to a C++ Bound object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Holon]{rchsm::C_Holon}} -> \code{C_Bound}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_Bound-new}{\code{C_Bound$new()}}
\item \href{#method-C_Bound-clone}{\code{C_Bound$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="finalize"><a href='../../rchsm/html/C_Object.html#method-C_Object-finalize'><code>rchsm::C_Object$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Holon" data-id="getValueString"><a href='../../rchsm/html/C_Holon.html#method-C_Holon-getValueString'><code>rchsm::C_Holon$getValueString()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Bound-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_Bound-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Bound$new(source, className = "Bound")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Source data for creating the bound.
Providing a character string will result in a new bound object
  being created in both C and R.
Providing an external pointer will result in a new bound object
  in R that is mapped to the C object referenced by the pointer.}

\item{\code{className}}{(Optional) Name of the class used in accessing c++ functions.
Default value is "Bound"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Bound-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_Bound-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Bound$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
