% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Cell.R
\name{C_Cell}
\alias{C_Cell}
\title{Cell Object}
\description{
An R6 class mapping to a C++ Cell object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Holon]{rchsm::C_Holon}} -> \code{C_Cell}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_Cell-new}{\code{C_Cell$new()}}
\item \href{#method-C_Cell-clone}{\code{C_Cell$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Holon" data-id="getValueString"><a href='../../rchsm/html/C_Holon.html#method-C_Holon-getValueString'><code>rchsm::C_Holon$getValueString()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Cell-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_Cell-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Cell$new(source, className = "Cell", regFinalizer = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Source data for creating the cell.
Providing a character string will result in a new cell object
  being created in both C and R.
Providing an external pointer will result in a new cell object
  in R that is mapped to the C object referenced by the pointer.}

\item{\code{className}}{(Optional) Name of the class used in accessing c++ functions.
Default value is "Cell"}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be 
registered (TRUE) or not (FALSE).
Defaults to TRUE.
Note this argument has no effect if the object is created from
an existing external pointer.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Cell-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_Cell-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Cell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
