% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Holon.R
\name{C_Holon}
\alias{C_Holon}
\title{CHSM Holon Object}
\description{
A holon in a composite Hierarchy State Machine model.

  Implementation is provided by C++ objects.
}
\section{Super class}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{C_Holon}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_Holon-new}{\code{C_Holon$new()}}
\item \href{#method-C_Holon-getValueString}{\code{C_Holon$getValueString()}}
\item \href{#method-C_Holon-clone}{\code{C_Holon$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="finalize"><a href='../../rchsm/html/C_Object.html#method-C_Object-finalize'><code>rchsm::C_Object$finalize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Holon-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_Holon-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Holon$new(className = "Holon", .external = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{className}}{(Optional) Name of the class used in accessing c++ functions.
Default value is "Holon"}

\item{\code{.external}}{(Optional) If an external pointer is provided instead of the default
NULL value, then the new R6 object will be associated with the 
existing C++ Object referenced by the pointer.}

\item{\code{...}}{Abstract arguments passed to the constructor of the super class}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Holon-getValueString"></a>}}
\if{latex}{\out{\hypertarget{method-C_Holon-getValueString}{}}}
\subsection{Method \code{getValueString()}}{
Constructs an object that is a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Holon$getValueString()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A single element character vector representing the contents of the
  holon.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Holon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_Holon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Holon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
