% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Model.R
\name{C_Model}
\alias{C_Model}
\title{CHSM Model Object}
\description{
A Composite Hierarchy State Machine model.

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_Variable]{rchsm::C_Variable}} -> \code{\link[rchsm:C_Holon]{rchsm::C_Holon}} -> \code{C_Model}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_Model-new}{\code{C_Model$new()}}
\item \href{#method-C_Model-createBound}{\code{C_Model$createBound()}}
\item \href{#method-C_Model-createCell}{\code{C_Model$createCell()}}
\item \href{#method-C_Model-createVariable}{\code{C_Model$createVariable()}}
\item \href{#method-C_Model-init}{\code{C_Model$init()}}
\item \href{#method-C_Model-installReporter}{\code{C_Model$installReporter()}}
\item \href{#method-C_Model-run}{\code{C_Model$run()}}
\item \href{#method-C_Model-clone}{\code{C_Model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Variable" data-id="getValueString"><a href='../../rchsm/html/C_Variable.html#method-C_Variable-getValueString'><code>rchsm::C_Variable$getValueString()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Holon" data-id="getVariablePointer"><a href='../../rchsm/html/C_Holon.html#method-C_Holon-getVariablePointer'><code>rchsm::C_Holon$getVariablePointer()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-new}{}}}
\subsection{Method \code{new()}}{
Constructs an object that is a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$new(
  depManager,
  solver,
  name = "Model",
  className = "Machine",
  regFinalizer = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{depManager}}{The dependency manager C object the machine should use}

\item{\code{solver}}{The solver C object the machine should use}

\item{\code{name}}{(Optional) Name of the model.
Default value is "Model".}

\item{\code{className}}{(Optional) Name of the class used in accessing c++ functions.
Default value is "Machine"}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be 
registered (TRUE) or not (FALSE).
Defaults to TRUE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-createBound"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-createBound}{}}}
\subsection{Method \code{createBound()}}{
Creates a new bound in the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$createBound(name, cellFrom, cellTo, holon = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the bound to be created}

\item{\code{cellFrom}}{R6 bound object for the from cell}

\item{\code{cellTo}}{R6 bound object for the to cell}

\item{\code{holon}}{(Optional) R6 holon object where the bound should be created.
Default NULL value results in creation of the bound in the model holon}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 bound object mapped to the C++ bound object that was created.
  The function will cause a fatal error in R if the c object returns
    an error message.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-createCell"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-createCell}{}}}
\subsection{Method \code{createCell()}}{
Creates a new cell in the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$createCell(name, holon = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the cell to be created}

\item{\code{holon}}{(Optional) R6 holon object where the cell should be created.
Default NULL value results in creation of the cell in the model holon}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new R6 cell object mapped to the C++ cell object that was created.
  The function will cause a fatal error in R if the c object returns
    an error message.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-createVariable"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-createVariable}{}}}
\subsection{Method \code{createVariable()}}{
Creates a new variable in the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$createVariable(name, value, holon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string with the name of the variable}

\item{\code{value}}{Value object of the new variable}

\item{\code{holon}}{Holon object containing the new variable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
External pointer to the variable that was created
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-init"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-init}{}}}
\subsection{Method \code{init()}}{
Initialize the state machine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$init()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-installReporter"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-installReporter}{}}}
\subsection{Method \code{installReporter()}}{
Install a reporter in the state machine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$installReporter(reporter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reporter}}{The reporter object to install}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-run"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-run}{}}}
\subsection{Method \code{run()}}{
Run the state machine simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
