% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Reporter.R
\name{C_ReporterTable}
\alias{C_ReporterTable}
\title{CHSM Table reporter object}
\description{
An R reference to a table reporter object

  Implementation is provided by C++ objects.
}
\section{Super classes}{
\code{\link[rchsm:C_Object]{rchsm::C_Object}} -> \code{\link[rchsm:C_ReporterInterval]{rchsm::C_ReporterInterval}} -> \code{C_ReporterTable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-C_ReporterTable-new}{\code{C_ReporterTable$new()}}
\item \href{#method-C_ReporterTable-getDataFrame}{\code{C_ReporterTable$getDataFrame()}}
\item \href{#method-C_ReporterTable-trackRate}{\code{C_ReporterTable$trackRate()}}
\item \href{#method-C_ReporterTable-trackVariable}{\code{C_ReporterTable$trackVariable()}}
\item \href{#method-C_ReporterTable-clone}{\code{C_ReporterTable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchsm" data-topic="C_Object" data-id="callFunction"><a href='../../rchsm/html/C_Object.html#method-C_Object-callFunction'><code>rchsm::C_Object$callFunction()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ReporterTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-C_ReporterTable-new}{}}}
\subsection{Method \code{new()}}{
Create an instance of the reporter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ReporterTable$new(
  interval,
  iterationVariable,
  className = "ReporterTable",
  regFinalizer = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interval}}{Integer defining the iteration interval for reporting}

\item{\code{iterationVariable}}{Variable with the iteration counter value}

\item{\code{className}}{(Optional) Character string with the name of the class.
Defaults to "ReporterTable".}

\item{\code{regFinalizer}}{(Optional) Logical to indicate if a finalizer should be registered (TRUE)
or not (FALSE).
Defaults to FALSE.
See super class C_Object for more information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ReporterTable-getDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-C_ReporterTable-getDataFrame}{}}}
\subsection{Method \code{getDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ReporterTable$getDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ReporterTable-trackRate"></a>}}
\if{latex}{\out{\hypertarget{method-C_ReporterTable-trackRate}{}}}
\subsection{Method \code{trackRate()}}{
Designate a rate variable with value(s) to be tracked by the reporter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ReporterTable$trackRate(variable, from = TRUE, to = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{Either an external pointer or an R6 object referring to the rate 
variable to be tracked.}

\item{\code{from}}{Logical TRUE if the rate value on the to side should be added, 
FALSE if not.}

\item{\code{to}}{Logical TRUE if the rate value on the from side should be added,
FALSE if not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ReporterTable-trackVariable"></a>}}
\if{latex}{\out{\hypertarget{method-C_ReporterTable-trackVariable}{}}}
\subsection{Method \code{trackVariable()}}{
Designate a variable with a value to be tracked by the reporter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ReporterTable$trackVariable(variable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variable}}{Either an external pointer or an R6 object referring to the variable
to be tracked.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_ReporterTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-C_ReporterTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{C_ReporterTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
